Version 1.1
GeoJSONs of the German federal motorways/Autobahnen numbers 21 (North to South and South to North), 25 (East to West and West to East), 7 (North to South), 27 (North to South and South to North) and 28 (East to West and West to East). Slip roads to the motorways are messy. Intended for use in Markus Bergelts GEOlayers 3 within Adobe After Effects.